Factor out method gathering.
